5.16 Модуляция сделано
